refactor(app): move some utils to `/profile`
