+++ app/profile/[user-id]/@profile/page.tsx
@@ -6,7 +6,7 @@ import { ACCESS_TOKEN } from '@app/api/constants'
 import { getUser } from '@app/api/fetchers'
 import { ProfileCard } from '@app/profile/components/profile'
 import { ProfilePageProps } from '@app/profile/types'
-import { validateUserId } from '@app/utils/user-id'
+import { validateUserId } from '@app/profile/utils/user-id'
 
 export default async function ProfileSubPage({ params }: ProfilePageProps) {
   const accessToken = cookies().get(ACCESS_TOKEN)?.value
