+++ app/profile/[user-id]/@top/@genres/page.tsx
@@ -4,11 +4,11 @@ import { cookies } from 'next/headers'
 import { ProfilePageProps } from '@app/profile/types'
 import { TIME_RANGE, USER_ID } from '@app/constants'
 import { ACCESS_TOKEN } from '@app/api/constants'
-import { validateTimeRange } from '@app/utils/time-range'
+import { validateTimeRange } from '@app/profile/utils/time-range'
 import { GenreChip, SeeMoreButton } from '@app/components/common'
 import { DefaultSection } from '@app/sections'
 import { getTopGenres } from '@app/api/fetchers'
-import { validateUserId } from '@app/utils/user-id'
+import { validateUserId } from '@app/profile/utils/user-id'
 
 export default async function ProfileTopGenresSubPage({
   searchParams,
